requirements.txt
numpy
scipy
Pillow
torch
torchvision
tqdm
pytorch-fid
huggingface_hub
onnxruntime-gpu
matplotlib